+++ .changeset/dirty-apricots-shout.md
@@ -1,0 +1,5 @@
+---
+"open-next": patch
+---
+
+Set `__NEXT_PRIVATE_PREBUNDLED_REACT` to use prebundled React
