+++ packages/open-next/src/adapters/server-adapter.ts
@@ -15,16 +15,14 @@ import { debug } from "./logger.js";
 import type { PublicFiles } from "../build.js";
 import { convertFrom, convertTo } from "./event-mapper.js";
 
+setNodeEnv();
+setNextjsServerWorkingDirectory();
 const nextDir = path.join(__dirname, ".next");
 const openNextDir = path.join(__dirname, ".open-next");
 const config = loadConfig(nextDir);
-
-setNodeEnv(config);
-
-setNextjsServerWorkingDirectory();
-
 const htmlPages = loadHtmlPages();
 const publicAssets = loadPublicAssets();
+setNextjsPrebundledReact(config);
 debug({ nextDir });
 
 // Create a NextServer
@@ -110,6 +108,14 @@ export async function handler(
 function setNextjsServerWorkingDirectory() {
   // WORKAROUND: Set `NextServer` working directory (AWS specific) — https://github.com/serverless-stack/open-next#workaround-set-nextserver-working-directory-aws-specific
   process.chdir(__dirname);
+}
+
+function setNextjsPrebundledReact(config: any) {
+  // WORKAROUND: Set `__NEXT_PRIVATE_PREBUNDLED_REACT` to use prebundled React — https://github.com/serverless-stack/open-next#workaround-set-__next_private_prebundled_react-to-use-prebundled-react
+  process.env.__NEXT_PRIVATE_PREBUNDLED_REACT = config.experimental
+    .serverActions
+    ? "experimental"
+    : "next";
 }
 
 function loadHtmlPages() {
